part 1 of feed list refactor
