+++ Modules/feed/Views/feed_delete_modal.php
@@ -1,0 +1,575 @@
+
+<!------------------------------------------------------------------------------------------------------------------------------------------------- -->
+<!-- FEED DELETE MODAL                                                                                                                             -->
+<!------------------------------------------------------------------------------------------------------------------------------------------------- -->
+<div id="feedDeleteModal" class="modal hide" tabindex="-1" role="dialog" aria-labelledby="feedDeleteModalLabel" aria-hidden="true" data-backdrop="static">
+    <div class="modal-header">
+        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">×</button>
+        <h3 id="feedDeleteModalLabel"><?php echo tr('Delete feed'); ?> 
+        <span id="feedDelete-message" class="label label-warning" data-default="<?php echo tr('Deleting a feed is permanent.'); ?>"><?php echo tr('Deleting a feed is permanent.'); ?></span>
+        </h3>
+    </div>
+    <div class="modal-body">
+        <div class="clearfix d-flex row">
+            <div id="clearContainer" class="span6">
+                <div style="min-height:12.1em; position:relative" class="well well-small">
+                    <h4 class="text-info"><?php echo tr('Clear') ?>:</h4>
+                    <p><?php echo tr('Empty feed of all data') ?></p>
+                    <button id="feedClear-confirm" class="btn btn-inverse" style="position:absolute;bottom:.8em"><?php echo tr('Clear Data'); ?>&hellip;</button>
+                </div>
+            </div>
+
+            <div id="trimContainer" class="span6">
+                <div class="well well-small">
+                    <h4 class="text-info"><?php echo tr('Trim') ?>:</h4>
+                    <p><?php echo tr('Empty feed data up to') ?>:</p>
+                    <div id="trim_start_time_container" class="control-group" style="margin-bottom:1.3em">
+                        <div class="controls">
+                            <div id="feed_trim_datetimepicker" class="input-append date" style="margin-bottom:0">
+                                <input id="trim_start_time" class="input-medium" data-format="dd/MM/yyyy hh:mm:ss" type="text" placeholder="dd/mm/yyyy hh:mm:ss">
+                                <span class="add-on"> <i data-time-icon="icon-time" data-date-icon="icon-calendar" class="icon-calendar"></i></span>
+                            </div>
+                            <div class="btn-group" style="margin-bottom:-4px">
+                                <button class="btn btn-mini active" title="<?php echo tr('Set to the start date') ?>" data-relative_time="start"><?php echo tr('Start') ?></button>
+                                <button class="btn btn-mini" title="<?php echo tr('One year ago') ?>" data-relative_time="-1y"><?php echo tr('- 1 year') ?></button>
+                                <button class="btn btn-mini" title="<?php echo tr('Two years ago') ?>" data-relative_time="-2y"><?php echo tr('- 2 year') ?></button>
+                                <button class="btn btn-mini" title="<?php echo tr('Set to the current date/time') ?>" data-relative_time="now"><?php echo tr('Now') ?></button>
+                            </div>
+                        </div>
+                    </div>
+                    <button id="feedTrim-confirm" class="btn btn-inverse"><?php echo tr('Trim Data'); ?>&hellip;</button>
+                </div>
+            </div>
+        </div>
+        
+        <div class="well well-small" style="margin-bottom:0">
+            <h4 class="text-info"><?php echo tr('Delete')?>: <span id="feedProcessList"></span></h4>
+            <p id="deleteFeedText"><?php echo tr('If you have Input Processlist processors that use this feed, after deleting it, review that process lists or they will be in error, freezing other Inputs. Also make sure no Dashboards use the deleted feed.'); ?></p>
+            <p id="deleteVirtualFeedText"><?php echo tr('This is a Virtual Feed, after deleting it, make sure no Dashboard continue to use the deleted feed.'); ?></p>
+            <button id="feedDelete-confirm" class="btn btn-danger"><?php echo tr('Delete feed permanently'); ?></button>
+        </div>
+    </div>
+    <div class="modal-footer">
+        <div id="feeds-to-delete" class="pull-left"></div>
+        <div id="feedDelete-loader" class="ajax-loader" style="display:none;"></div>
+        <button class="btn" data-dismiss="modal" aria-hidden="true"><?php echo tr('Close'); ?></button>
+    </div>
+</div>
+<script>
+
+
+// ---------------------------------------------------------------------------------------------
+// DELETE FEED
+// ---------------------------------------------------------------------------------------------
+
+/**
+ * getFeedProcess
+ * 
+ * Scans all input processes and identifies which processes are linked to feeds.
+ * Returns an object mapping feed IDs to their associated input, process definition, and feed ID.
+ * Used to determine which feeds are referenced by input process lists for safe deletion and management.
+ *
+ * @returns {Object} feedInputs - Map of feed IDs to input/process details.
+ */
+function getFeedProcess() {
+    const feedInputs = {};
+
+    // 1. Build a map of processes that are linked to feeds
+    const feedProcesses = {};
+    for (const key in process_vue.processes_by_key) {
+        const proc = process_vue.processes_by_key[key];
+
+        // Cycle through args, check for feed args
+        for (let i = 0; i < proc.args.length; i++) {
+            const arg = proc.args[i];
+            if (arg.type === 2) {
+                // Could filter for set engines field here but think it's useful to flag all feed linked processes
+                feedProcesses[key] = proc;
+                break; // No need to check further args
+            }
+        }
+    }
+
+    // 2. For each input, decode its processList and check for feed linked processes
+    for (const input of Object.values(process_vue.inputs)) {
+        if (!input.processList.length) continue;
+        const decodedList = process_api.decode(input.processList);
+
+        decodedList.forEach(procItem => {
+            const procDef = feedProcesses[procItem.fn];
+            if (!procDef) return;
+            procDef.args.forEach((arg, idx) => {
+                if (arg.type === 2) {
+                    const feedid = procItem.args[idx];
+                    feedInputs[feedid] = {
+                        input: {
+                            nodeid: input.nodeid,
+                            name: input.name,
+                            inputid: input.id,
+                            // processList: decodedList (Not currently used)
+                        },
+                        process: procDef,
+                        feedid: feedid
+                    };
+                }
+            });
+        });
+    }
+
+    return feedInputs;
+}
+
+/**
+ * output what feeds have been selected in the overlay modal box
+ *
+ * @return void
+ */
+function showSelectedFeeds(feed_inputs) {
+    let total_selected = 0;
+    let total_input_processes_linked = 0;
+    let total_virtual_feed_processes_linked = 0;
+    let feedListShort = '';
+
+    for (const feedid in selected_feeds) {
+        if (!selected_feeds[feedid]) continue;
+        total_selected++;
+        const feed = feeds[feedid];
+        feedListShort += `[${feedid}] ${feed.tag}:${feed.name}, `;
+
+        // Virtual feed: count its processList items
+        if (parseInt(feed.engine) === 7 && feed.processList != "") {
+            total_virtual_feed_processes_linked += process_api.decode(feed.processList).length;
+        }
+        // Non-virtual: count input processes referencing this feed
+        else if (feed_inputs[feedid]) {
+            if (Array.isArray(feed_inputs[feedid])) {
+                total_input_processes_linked += feed_inputs[feedid].length;
+            } else {
+                total_input_processes_linked += 1;
+            }
+        }
+    }
+
+    // Remove trailing comma
+    if (feedListShort.endsWith(', ')) feedListShort = feedListShort.slice(0, -2);
+
+    // Build summary
+    let total_summary = '<div id="deleteFeedModalSelectedItems">';
+    let feedProcessList = '';
+
+    if (total_selected === 1) {
+        total_summary += `<h5>${feedListShort}</h5>`;
+    } else {
+        total_summary += `<h5 title="${feedListShort}"><?php echo tr('%s Feeds selected') ?> <i class="icon icon-question-sign"></i></h5>`.replace('%s', total_selected);
+    }
+
+    // Compose the combined message
+    if (total_input_processes_linked > 0 || total_virtual_feed_processes_linked > 0) {
+        let msg = '';
+        if (total_input_processes_linked > 0) {
+            msg += total_input_processes_linked + ' <?php echo tr("input processes") ?>';
+        }
+        if (total_virtual_feed_processes_linked > 0) {
+            if (msg.length > 0) msg += ' and ';
+            msg += total_virtual_feed_processes_linked + ' <?php echo tr("virtual feed processes") ?>';
+        }
+        msg += total_selected === 1
+            ? ' <?php echo tr("associated with this feed") ?>'
+            : ' <?php echo tr("associated with these feeds") ?>';
+
+        feedProcessList = `<span class="badge badge-default" style="padding-left:4px;margin-right:6px"><i class="icon icon-white icon-exclamation-sign"></i> ${msg}</span>`;
+    }
+
+    total_summary += '</div>';
+
+    $("#feeds-to-delete").html(total_summary);
+    $("#feedProcessList").html(feedProcessList);
+}
+
+/**
+ * show the trim start time in the date time picker and input field
+ * 
+ * will also highlight a button if it matches the currently selected timestamp
+ *
+ * @param int start_time unix timestamp (seconds)
+ */
+function showFeedStartDate(start_time){
+    let startDate = start_time==0 ? new Date() : new Date(start_time*1000);
+    $datetimepicker = $('#feed_trim_datetimepicker');
+    $datetimepicker
+        .datetimepicker({startDate: startDate}) // restrict calendar selection to the start time
+        .datetimepicker('setValue', startDate) // set the date/time picker to the start time
+        .on('changeDate', function(event){
+            // mark any matching buttons as active
+            $('[data-relative_time]').each(function(i,elem){
+                if ($(elem).data('startdate') != event.date) {
+                    $(this).removeClass('active')
+                }
+            });
+        });
+}
+
+/**
+ * Initialises the different events to enable the "relative date" selections below the date/time picker
+ * 
+ * Set the data property of each button to store correct Date() for each button. Each button must have a 
+ * "data-relative_time" attribute with one of the following values:-
+ *   "-2y", "-1y", "start" or "now" (default)
+ *
+ * Each button shows a formatted date in the input field and also sets the date time picker to the relevant position
+ * @param int start_time the earliest possible timestamp for all the selected feeds - does not allow trimming beyond this point
+ *
+ * @return void
+ */
+function initRelativeStartDateButtons(start_time){
+    let startDate = start_time>0 ? new Date(start_time*1000) : new Date();
+
+    $('[data-relative_time]').each(function(i,v){
+        $btn = $(this);
+        // add more cases here for additional options (and also data-relative_time='xyz' in the html)
+        // returns function so that the dates are calculated to when the user clicks the buttons
+        switch ($btn.data('relative_time')) {
+        case '-2y':
+            relativeTime = (function(){ now = new Date(); return new Date(now.getFullYear()-2,now.getMonth(),now.getDate(),now.getHours(),now.getMinutes(),now.getSeconds(),now.getMilliseconds()) });
+            break;
+        case '-1y':
+            relativeTime = (function(){ now = new Date(); return new Date(now.getFullYear()-1,now.getMonth(),now.getDate(),now.getHours(),now.getMinutes(),now.getSeconds(),now.getMilliseconds()) });
+            break;
+        case 'start':
+            relativeTime = startDate;
+            break;
+        default:
+            relativeTime = new Date();
+        }
+        relativeTime = typeof relativeTime === 'function' ? relativeTime() : relativeTime;
+        // set the timestamp as a data property of the button so that it can be referenced on click
+        $btn.data('startdate', relativeTime.valueOf() );
+        // make sure the calculated date is not beyond the start date
+        if (relativeTime < startDate) {
+            $btn.hide() // hide button date is beyond start date
+            $btn.css({'font-style':'italic', color:'#9a9eaa'});
+            $btn.attr('title',$btn.attr('title')+' - [<?php echo tr('Out of range')?>]');
+        }
+    })
+    // open date picker on input focus
+    $('#trim_start_time').on('focus', function(event){ $datetimepicker.datetimepicker('show') });
+    
+    // alter the trim date / time picker on button presses
+    $('[data-relative_time]').click(function(event){
+        event.preventDefault();
+        $btn = $(this);
+        $btn.addClass('active').siblings().removeClass('active');
+        $input = $('#trim_start_time');
+        // get starttime from button's data
+        date = new Date($btn.data('startdate'));
+        // restrict selection to the earliest possible date
+        if (date < startDate) {
+            date = startDate;
+        }
+        // rebuild the date string from the new date object
+        Y = date.getFullYear();
+        m = (date.getMonth()+1).pad(2);
+        d = date.getDate().pad(2);
+        h = date.getHours().pad(2);
+        i = date.getMinutes().pad(2);
+        s = date.getSeconds().pad(2);
+        
+        // show date in input field - DD/MM/YYYY HH:MM:SS
+        newDateString = [[d,m,Y].join('/'),[h,i,s].join(':')].join(' ');
+        $input.val(newDateString);
+    });
+}
+
+/**
+ * compares all the selected feed start_times to see which is the best suited for the group 
+ * @return int start_time timestamp
+ */
+function getEarliestStartTime() { 
+    let start_time = 0;
+    for (var feedid in selected_feeds) {
+        if (selected_feeds[feedid] == true) {
+            // record the earliest possible start_time for all the selected feeds
+            start_time = feeds[feedid].start_time > start_time ? feeds[feedid].start_time : start_time;
+        }
+    }
+    return start_time;
+}
+
+/**
+ * mark button as selected if chosen date in date/time picker matches
+ * jQuery Event handler for datetime picker's changeDate event
+ */
+$('#feed_trim_datetimepicker').on('changeDate',function(event){
+    $('[data-relative_time]').each(function(){
+        $btn = $(this);
+        if ($btn.data('startdate') == event.date.valueOf()) {
+            $btn.addClass('active').siblings().removeClass('active');
+        }
+    })
+});
+
+/**
+ * returns true if trim function available for all of the selected feed engine types
+ *
+ * @return boolean
+ */
+function isSelectionValidForTrim(){
+    /*
+        const MYSQL = 0;
+        const TIMESTORE = 1;     // Depreciated
+        const PHPTIMESERIES = 2;
+        const GRAPHITE = 3;      // Not included in core
+        const PHPTIMESTORE = 4;  // Depreciated
+        const PHPFINA = 5;
+        const VIRTUALFEED = 7;   // Virtual feed, on demand post processing
+        const MYSQLMEMORY = 8;   // Mysql with MEMORY tables on RAM. All data is lost on shutdown 
+        const REDISBUFFER = 9;   // (internal use only) Redis Read/Write buffer, for low write mode
+        const CASSANDRA = 10;    // Cassandra
+    */
+    let allowed_engines = [0,2,5,8] // array of allowed storage engines
+    for (var feedid in selected_feeds) {
+        engineid = parseInt(feeds[feedid].engine); // convert string to number
+        // if feed selected and engineid is NOT found in allowed_engines
+        if (selected_feeds[feedid] == true && !isNaN(engineid) && allowed_engines.indexOf(engineid) == -1) {
+            return false;
+        }
+    }
+    return true;
+}
+
+/**
+ * display a message to the user in the delete feed modal
+ *
+ * restores the original message after delay
+ *
+ * @param string message text to show to user
+ */
+function updateFeedDeleteModalMessage(response){
+    let message = response.message;
+    let success = response.success;
+    let $msg = $('#feedDelete-message');
+    let cssClassName = success ? 'label-success' : 'label-important';
+
+    $msg.stop().fadeOut(function(){
+        $(this).text(message).removeClass('label-warning').addClass(cssClassName).fadeIn();
+    });
+    setTimeout(function(){
+        $msg.stop().fadeOut(function(){
+            $msg.text($msg.data('default')).removeClass(cssClassName).addClass('label-warning').fadeIn();
+        })
+    }, 3800);
+}
+
+/**
+ * function call queue - clears previous interval if interrupted
+ */
+var updater;
+function updaterStart(func, interval){
+    clearInterval(updater);
+    updater = null;
+    if (interval > 0) updater = setInterval(func, interval);
+}
+
+/**
+ * Enables/Disables the feed trim() feature based on selected feeds
+ *
+ * @return void
+ */
+function initTrim(){
+    // get the most suitable start_time for all selected feeds
+    if (isSelectionValidForTrim()) {
+        let start_time = getEarliestStartTime()
+        enableTrim(start_time);
+    } else {
+        disableTrim();
+    }
+}
+
+/**
+ * Allows feed(s) to be trimmed to a new start_date
+ *
+ * @param int start_time new timestamp to trim to
+ * @return void
+ */
+function enableTrim(start_time){
+    // populate the trim() date input with the feed's current start date
+    showFeedStartDate(start_time);
+    // make buttons under the trim date input react on click
+    initRelativeStartDateButtons(start_time);
+
+    // remove any styling the disableTrim() function created
+    $('#trimContainer').attr('title','').removeClass('muted')//.show()
+        .find('h4').addClass('text-info').removeClass('muted').end()
+        .find('button,input').removeClass('disabled')
+        .find('input').val('');
+    
+    // enable the confirm trim button
+    $('#feedTrim-confirm')
+        .unbind('click')
+        .click(function(){
+            $modal = $('#feedDeleteModal');
+            let $input = $modal.find("#trim_start_time");
+            let input_date_string = $input.val();
+            // dont submit if nothing selected
+            // convert uk dd/mm/yyyy h:m:s to RFC2822 date
+            let start_date = new Date(input_date_string.replace( /(\d{2})\/(\d{2})\/(\d{4}) (\d{2}):(\d{2}):(\d{2})/, "$3-$2-$1T$4:$5:$6"));
+            let isValidDate = !isNaN(start_date.getTime()) && input_date_string != "";
+            // exit if supplied date not valid
+            if (!isValidDate) {
+                $('#trim_start_time_container').addClass('error');
+                $input.focus();
+                return false;
+            }else{
+                if(confirm("<?php echo tr('This is a new feature. Consider backing up your data before you continue. OK to continue?') ?>") == true) {
+                    $('#trim_start_time_container').removeClass('error');
+                    // set to seconds from milliseconds
+                    let start_time = start_date.getTime()/1000;
+                    $("#feedDelete-loader").fadeIn();
+                    // run the trim() function on all the selected feeds
+                    for (let feedid in selected_feeds) {
+                        if (selected_feeds[feedid]) {
+                            let response = feed.trim(feedid, start_time);
+                            updateFeedDeleteModalMessage(response);
+                            if (!response.success) {
+                                break;
+                            }
+                        }
+                    }
+                    $("#feedDelete-loader").stop().fadeOut();
+                    update_feed_list();
+                    updaterStart(update_feed_list, 5000);
+                }
+            }
+        });
+}
+
+/**
+ * hide the trim feature
+ *
+ * @return void
+ */
+function disableTrim(){
+    $('#trimContainer').attr('title','<?php echo tr('"Trim" not available for this storage engine') ?>').addClass('muted')//.hide()
+        .find('h4').removeClass('text-info').addClass('muted').end()
+        .find('button,input').addClass('disabled')
+        .find('input').val('');
+    $('#feedTrim-confirm').unbind('click'); // remove previous click event (if it exists)
+}
+
+/**
+ * trigger off the modal overlay to display delete options
+ * 
+ * jQuery Event handler for the delete feed button
+ * also shows items selected as well as a processlist warning
+ */
+$(".feed-delete").click(function(){
+    $('#feedDeleteModal #deleteFeedText').show();
+    $('#feedDeleteModal #deleteVirtualFeedText').hide();
+    $('#feedDeleteModal').modal('show'); //show the delete modal
+
+    // get the list of input processlists that write to feeds
+    let feed_processes = getFeedProcess();
+    let selected_feeds_inputs = {};
+    for (i in selected_feeds){
+        // if a selected feed has an associated process id then save it into an array
+        if (selected_feeds[i] && typeof feed_processes[i] != 'undefined') {
+            selected_feeds_inputs[i] = feed_processes[i];
+        }
+    }
+
+    // show the selected feeds and any associated processList
+    showSelectedFeeds(selected_feeds_inputs);
+
+    initTrim();
+    initClear();
+});
+
+function isSelectionValidForClear(){
+    /*
+        const MYSQL = 0;
+        const TIMESTORE = 1;     // Depreciated
+        const PHPTIMESERIES = 2;
+        const GRAPHITE = 3;      // Not included in core
+        const PHPTIMESTORE = 4;  // Depreciated
+        const PHPFINA = 5;
+        const VIRTUALFEED = 7;   // Virtual feed, on demand post processing
+        const MYSQLMEMORY = 8;   // Mysql with MEMORY tables on RAM. All data is lost on shutdown 
+        const REDISBUFFER = 9;   // (internal use only) Redis Read/Write buffer, for low write mode
+        const CASSANDRA = 10;    // Cassandra
+    */
+    let allowed_engines = [0,2,5,8]; // array of allowed storage engines 
+    for (var feedid in selected_feeds) {
+        engineid = parseInt(feeds[feedid].engine); // convert string to number
+        // if feed selected and engineid is NOT found in allowed_engines
+        if (selected_feeds[feedid] == true && !isNaN(engineid) && allowed_engines.indexOf(engineid) == -1) {
+            return false;
+        }
+    }
+    return true;
+}
+
+function initClear(){
+    // get the most suitable start_time for all selected feeds
+    if (isSelectionValidForClear()) {
+        enableClear();
+    } else {
+        disableClear();
+    }
+}
+
+function enableClear(){
+    // remove any disable styling
+    $('#clearContainer').attr('title','').removeClass('muted')//.show()
+        .find('h4').addClass('text-info').removeClass('muted').end()
+        .find('button').removeClass('disabled');
+
+    $("#feedClear-confirm")
+        .unbind('click')
+        .click(function(){
+            if( confirm("<?php echo tr('Are you sure you want to delete all the feeds data?') ?>") == true ){
+                $modal = $('#feedDeleteModal');
+                $("#feedDelete-loader").fadeIn();
+
+                for (let feedid in selected_feeds) {
+                    if (selected_feeds[feedid]) {
+                        let response = feed.clear(feedid);
+                        updateFeedDeleteModalMessage(response);
+                        if (!response.success) {
+                            break;
+                        }
+                    }
+                }
+                $("#feedDelete-loader").stop().fadeOut();
+                update_feed_list();
+                updaterStart(update_feed_list, 5000);
+            }
+        });
+}
+
+function disableClear(){
+    $("#feedClear-confirm").unbind();
+
+    $('#clearContainer').attr('title','<?php echo tr('"Clear" not available for this storage engine') ?>').addClass('muted')//.hide()
+        .find('h4').removeClass('text-info').addClass('muted').end()
+        .find('button').addClass('disabled');
+}
+
+$("#feedDelete-confirm").click(function(){
+    if( confirm("<?php echo tr('Are you sure you want to delete?') ?>") == true) {
+        for (let feedid in selected_feeds) {
+            if (selected_feeds[feedid]) {
+                let response = feed.remove(feedid);
+                response = response ? response : {success:true, message: '<?php echo tr("Feeds Deleted") ?>'};
+                updateFeedDeleteModalMessage(response);
+            }
+        }
+        
+        setTimeout(function() {
+            update_feed_list();
+            updaterStart(update_feed_list, 5000);
+            $('#feedDeleteModal').modal('hide');
+            feed_selection();
+        }, 5000);
+    }
+});
+</script>
